Cleanup classes Remove custom attribute
