+++ src/X.Web.Sitemap/ChangeFrequency.cs
@@ -1,4 +1,5 @@
-﻿using System.Xml.Serialization;
+﻿using System;
+using System.Xml.Serialization;
 using JetBrains.Annotations;
 
 namespace X.Web.Sitemap;
