+++ src/X.Web.Sitemap/FileSystemWrapper.cs
@@ -1,7 +1,16 @@
 ﻿using System.IO;
 using System.Threading.Tasks;
+using JetBrains.Annotations;
 
 namespace X.Web.Sitemap;
+
+[PublicAPI]
+internal interface IFileSystemWrapper
+{
+    FileInfo WriteFile(string xml, string path);
+        
+    Task<FileInfo> WriteFileAsync(string xml, string path);
+}
 
 internal class FileSystemWrapper : IFileSystemWrapper
 {
