--- src/X.Web.Sitemap/SerializableAttribute.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using System.Text;
-
-namespace X.Web.Sitemap;
-
-public class SerializableAttribute : Attribute
-{
-}
